requirements.txt
numpy>=1.24.0
pandas>=2.0.0
numba>=0.57.0
cupy-cuda12x>=12.0.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
pytest>=7.3.0
